chore: split webtools/trustchain into components
