+++ apps/web-tools/trustchain/components/Input.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Input = styled.input`
+  padding: 0.8em;
+  width: 100%;
+  flex: 1;
+`;
