+++ apps/web-tools/trustchain/device.ts
@@ -1,0 +1,7 @@
+import { withDevice } from "@ledgerhq/live-common/hw/deviceAccess";
+import { from, lastValueFrom } from "rxjs";
+import Transport from "@ledgerhq/hw-transport";
+
+export function runWithDevice<T>(fn: (transport: Transport) => Promise<T>): Promise<T> {
+  return lastValueFrom(withDevice("webhid")(transport => from(fn(transport))));
+}
